# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/properties/{propertyId}/tenants/{tenantUserId}/block": {
      "post": {
        "operationId": "PropertyTenantsController_blockTenant",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bloque un locataire pour ce bien précis (justification obligatoire)",
        "tags": ["Tenants"]
      }
    },
    "/properties/{propertyId}/tenants/history": {
      "get": {
        "operationId": "PropertyTenantsController_getHistory",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historique des baux (passés et présent) de ce bien",
        "tags": ["Tenants"]
      }
    },
...
    "/leases/{id}/terminate": {
      "post": {
        "operationId": "LeasesController_terminate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminateLeaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Résilie un bail — libère le bien (VACANT), historique conservé",
        "tags": ["Leases"]
      }
    },
    "/leases/{id}/schedule": {
      "get": {
        "operationId": "LeasesController_getSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Calendrier complet des échéances d'un bail",
        "tags": ["Leases"]
      }
    },
...
    "/payment-declarations": {
      "post": {
        "description": "Réservé au locataire connecté, sur son propre bail. Justificatif (`proof`) optionnel.",
        "operationId": "PaymentDeclarationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDeclarationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Déclare avoir payé une échéance — passe en attente de confirmation",
        "tags": ["Payment Declarations"]
      }
    },
    "/payment-declarations/{id}/proof": {
      "get": {
        "operationId": "PaymentDeclarationsController_getProofUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retourne une URL signée pour consulter la preuve de paiement",
        "tags": ["Payment Declarations"]
      }
    },
    "/payment-declarations/{id}": {
      "patch": {
        "description": "Justificatif (`proof`) optionnel — remplace le précédent si fourni.",
        "operationId": "PaymentDeclarationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDeclarationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifie une déclaration tant qu’elle est en attente de confirmation",
        "tags": ["Payment Declarations"]
      },
      "delete": {
        "operationId": "PaymentDeclarationsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Annule une déclaration tant qu’elle est en attente de confirmation",
        "tags": ["Payment Declarations"]
      }
    },
...
    "/dashboard/manager/summary": {
      "get": {
        "operationId": "DashboardManagerController_getSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Synthèse des biens sous mandat actif, répartition par statut, nombre de propriétaires mandants",
        "tags": ["Dashboard"]
      }
    },
    "/dashboard/manager/revenue": {
      "get": {
        "operationId": "DashboardManagerController_getRevenue",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "MONTH",
              "enum": ["MONTH", "YEAR"],
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "description": "1-12, ignoré si period=YEAR",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "description": "Restreint le volet \"biens sous mandat\" à ce propriétaire mandant",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Encaissements de la période (mois/année), comparaison à la période précédente, ventilation biens propres/sous mandat",
        "tags": ["Dashboard"]
      }
    },
    "/dashboard/manager/alerts": {
      "get": {
        "operationId": "DashboardManagerController_getAlerts",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ALL",
              "enum": ["MANAGED", "OWNED", "ALL"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Échéances impayées, baux expirant sous 30 jours, déclarations en attente de confirmation",
        "tags": ["Dashboard"]
      }
    },
    "/dashboard/manager/owners": {
      "get": {
        "operationId": "DashboardManagerController_getOwners",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Propriétaires mandants actuels, avec le nombre de biens confiés à chacun",
        "tags": ["Dashboard"]
      }
    },
    "/dashboard/manager/upcoming-payments": {
      "get": {
        "operationId": "DashboardManagerController_getUpcomingPayments",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ALL",
              "enum": ["MANAGED", "OWNED", "ALL"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Prochaines échéances à venir, triées par date",
        "tags": ["Dashboard"]
      }
    },
    "/dashboard/manager/properties/{id}/performance": {
      "get": {
        "operationId": "DashboardManagerController_getPropertyPerformance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Taux de recouvrement à temps sur les 6 derniers mois pour un bien donné",
        "tags": ["Dashboard"]
      }
    },
...
  },
  "info": {
    "title": "WARAH API",
    "description": "API de gestion locative WARAH — marché togolais",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BlockTenantDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Justification obligatoire du blocage"
          }
        },
        "required": ["reason"]
      },
      "TerminateLeaseDto": {
        "type": "object",
        "properties": {
          "terminationReason": {
            "type": "string"
          }
        }
      },
...
      "CreatePaymentDeclarationDto": {
        "type": "object",
        "properties": {
          "scheduleEntryId": {
            "type": "string"
          },
          "declaredAmount": {
            "type": "number"
          },
          "declaredAt": {
            "type": "string",
            "description": "Date effective du paiement (ISO 8601)"
          },
          "declaredMethod": {
            "type": "string",
            "enum": ["CASH", "BANK_TRANSFER"]
          },
          "note": {
            "type": "string"
          }
        },
        "required": ["scheduleEntryId", "declaredAmount", "declaredAt", "declaredMethod"]
      },
      "UpdatePaymentDeclarationDto": {
        "type": "object",
        "properties": {
          "declaredAmount": {
            "type": "number"
          },
          "declaredAt": {
            "type": "string",
            "description": "Date effective du paiement (ISO 8601)"
          },
          "declaredMethod": {
            "type": "string",
            "enum": ["CASH", "BANK_TRANSFER"]
          },
          "note": {
            "type": "string"
          }
        }
      },
...
    }
  }
}